ub: Fix bugs in urma kernel and user lib
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I8J45G
CVE: NA

--------------------------------

Fix bugs in urma kernel and user lib
- Rollback process vtp entry status fix
- Fix vf and pf eid filling bug
- Fix the rebuild chain problem in rc mode
- Add support for query link rebuild finish
- opt create vtp contex, avoid frequent hmap find operation
- Fix that problem that cannot find pf_dev on vm open
- Fix the problem that upi format is not stable OPEN
- DFX bugfix, mod utp_id from u8 to u32, RC mode use tpgn in show res
- Fixed the problem of link building failure and TP resource leakage
- Fix the problem that the query result of urma_admin show is wrong
- fix slice cfg will not cfg to driver
- optimize create vtp context, avoid duplicate code
- Fixed the bug where no exception occurred when illegal commands entered
by urma admin
- Fix urma_sub and urma_kmod_sub codecheck
- Fix two problems: 1. dip table wrong print 2.use after free OPEN
- Fix the bug of modify tp attr
- fix tpsa_tp_flag not consistent with ubcore_tp_cfg_flag
- Bond event monitoring OPEN
- Namespace: assign eid for ns
- Encapsulation of hashmap find, destroy and insert operation
- Fix the problem that sip_table add doesn't check the number of param OPEN
- Adapt to alpha version chain negotiation

+++ drivers/ub/urma/ubcore/ubcore_umem.c
@@ -147,7 +147,7 @@ static int umem_dma_map(struct ubcore_umem *umem, uint64_t npages, unsigned long
 {
 	int ret;
 
-	ret = dma_map_sg_attrs(umem->ub_dev->dma_dev, umem->sg_head.sgl, npages,
+	ret = dma_map_sg_attrs(umem->ub_dev->dma_dev, umem->sg_head.sgl, (int)npages,
 		DMA_BIDIRECTIONAL, dma_attrs);
 	if (ret == 0) {
 		ubcore_log_err("Dma map failed, ret: %d\n", ret);
@@ -250,7 +250,7 @@ void ubcore_umem_release(struct ubcore_umem *umem)
 		return;
 
 	npages = umem_cal_npages(umem->va, umem->length);
-	dma_unmap_sg(umem->ub_dev->dma_dev, umem->sg_head.sgl, umem->nmap, DMA_BIDIRECTIONAL);
+	dma_unmap_sg(umem->ub_dev->dma_dev, umem->sg_head.sgl, (int)umem->nmap, DMA_BIDIRECTIONAL);
 	umem_free_sgt(umem);
 	umem_atomic_sub(npages, umem->owning_mm);
 	mmdrop(umem->owning_mm);
